Changed the main Slimbox API. Now the slimbox() method must be called on a link element or a set of link elements to enable the Slimbox function on it/them. The options object can be passed to this method.
Calling the slimbox() method on a set of link elements automatically creates a Slimbox gallery including all the images of the set.

The auto-activation of all links with the "lightbox" rel attribute on page load is now completely optional. To disable it, the last block of code must be commented or removed.

The code size is now 4.5kb which is 1kb less than Slimbox 1.41, even with the added features (yay!)